# apps/api/src/products/products.controller.ts
import { Body, Controller, Delete, Get, Param, Patch, Post, Query, Req, UseGuards } from '@nestjs/common';
import { Resource } from '@prisma/client';
import { Request } from 'express';
import { JwtAuthGuard } from '../auth/guards/jwt-auth.guard';
import { PermissionsGuard } from '../permissions/guards/permissions.guard';
import { RequirePermission } from '../permissions/decorators/require-permissions.decorator';
import { CreateProductDto } from './dto/create-product.dto';
import { ProductListQueryDto } from './dto/product-query.dto';
import { UpdateProductDto } from './dto/update-product.dto';
import { ProductsService } from './products.service';
import { ActiveUser } from '../auth/decorators/active-user.decorator';
import { ActiveUserData } from '../auth/types/active-user-data';

@UseGuards(JwtAuthGuard, PermissionsGuard)
@Controller('products')
export class ProductsController {
  constructor(private readonly productsService: ProductsService) {}

  @Post()
  @RequirePermission(Resource.products, 'write')
  async create(
    @Body() createProductDto: CreateProductDto,
    @ActiveUser() activeUser: ActiveUserData,
    @Req() request: Request,
  ) {
    return this.productsService.create(createProductDto, this.buildAuditContext(activeUser, request));
  }

  @Get()
  @RequirePermission(Resource.products)
  async findAll(@Query() query: ProductListQueryDto) {
    const page = query.page ?? 1;
    const size = query.size ?? 20;

    const result = await this.productsService.findAll(query);

    return {
      data: result.data,
      page: {
        page,
        size,
        total: result.total,
      },
    };
  }

  @Get(':id')
  @RequirePermission(Resource.products)
  async findOne(@Param('id') id: string) {
    return this.productsService.findOne(id);
  }

  @Patch(':id')
  @RequirePermission(Resource.products, 'write')
  async update(
    @Param('id') id: string,
    @Body() updateProductDto: UpdateProductDto,
    @ActiveUser() activeUser: ActiveUserData,
    @Req() request: Request,
  ) {
    return this.productsService.update(id, updateProductDto, this.buildAuditContext(activeUser, request));
  }

  @Delete(':id')
  @RequirePermission(Resource.products, 'write')
  async remove(@Param('id') id: string, @ActiveUser() activeUser: ActiveUserData, @Req() request: Request) {
    await this.productsService.remove(id, this.buildAuditContext(activeUser, request));
    return { success: true };
  }

  private buildAuditContext(activeUser: ActiveUserData | undefined, request: Request) {
    return {
      actor: activeUser,
      ip: request.ip ?? request.socket?.remoteAddress,
      userAgent: request.get('user-agent') ?? undefined,
    };
  }
}