# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Inventory App API
  version: 1.0.0
  summary: Internal inventory management service
  description: >
    Inventory management API covering authentication, products, stock flows, alerts,
    and configuration. All endpoints (except login) require a bearer token issued to
    internal users.
servers:
  - url: /api/v1
    description: Primary API entrypoint
...
security:
  - bearerAuth: []
paths:
...
  /products:
    get:
      tags: [Products]
      summary: 제품 목록
      operationId: listProducts
      parameters:
        - name: search
          in: query
          description: 코드 또는 이름 검색어
          schema:
            type: string
        - name: status
          in: query
          description: 재고 상태 필터
          schema:
            type: string
            enum: [normal, warn, low]
        - name: disabled
          in: query
          description: 사용 중지 여부 필터 (true/false)
          schema:
            type: boolean
        - name: includeDisabled
          in: query
          description: 사용 중지 제품을 목록에 포함할지 여부
          schema:
            type: boolean
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: 제품 목록 응답
          content:
            application/json:
              schema:
                type: object
                required: [data, page]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  page:
                    $ref: '#/components/schemas/Page'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Products]
      summary: 제품 생성
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, name]
              properties:
                code:
                  type: string
                name:
                  type: string
                description:
                  type: string
                safetyStock:
                  type: integer
                  minimum: 0
                disabled:
                  type: boolean
                  default: 0
      responses:
        '201':
          description: 제품 생성 완료
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /products/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: 제품 ID
        schema:
          type: string
    get:
      tags: [Products]
      summary: 제품 상세 조회
      operationId: getProduct
      responses:
        '200':
          description: 제품 상세
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags: [Products]
      summary: 제품 수정
      operationId: updateProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                  nullable: true
                safetyStock:
                  type: integer
                  minimum: 0
      responses:
        '200':
          description: 제품 수정 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags: [Products]
      summary: 제품 삭제
      operationId: deleteProduct
      responses:
        '204':
          description: 제품 삭제 완료
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /inbounds:
    get:
      tags: [StockFlows]
      summary: 입고 목록
      operationId: listInbounds
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: productId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 입고 목록
          content:
            application/json:
              schema:
                type: object
                required: [data, page]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Inbound'
                  page:
                    $ref: '#/components/schemas/Page'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [StockFlows]
      summary: 입고 등록
      operationId: createInbound
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Inbound'
      responses:
        '201':
          description: 입고 등록 완료
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /inbounds/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [StockFlows]
      summary: 입고 상세 조회
      operationId: getInbound
      responses:
        '200':
          description: 입고 상세
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inbound'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags: [StockFlows]
      summary: 입고 수정
      operationId: updateInbound
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productId:
                  type: string
                quantity:
                  type: integer
                  minimum: 1
                dateIn:
                  type: string
                  format: date
                note:
                  type: string
      responses:
        '200':
          description: 입고 수정 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inbound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags: [StockFlows]
      summary: 입고 삭제
      operationId: deleteInbound
      responses:
        '200':
          description: 입고 삭제 완료
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /outbounds:
    get:
      tags: [StockFlows]
      summary: 출고 목록
      operationId: listOutbounds
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: productId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 출고 목록
          content:
            application/json:
              schema:
                type: object
                required: [data, page]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Outbound'
                  page:
                    $ref: '#/components/schemas/Page'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [StockFlows]
      summary: 출고 등록
      operationId: createOutbound
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Outbound'
      responses:
        '201':
          description: 출고 등록 완료
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /outbounds/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [StockFlows]
      summary: 출고 상세 조회
      operationId: getOutbound
      responses:
        '200':
          description: 출고 상세
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Outbound'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags: [StockFlows]
      summary: 출고 수정
      operationId: updateOutbound
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productId:
                  type: string
                quantity:
                  type: integer
                  minimum: 1
                dateOut:
                  type: string
                  format: date
                note:
                  type: string
      responses:
        '200':
          description: 출고 수정 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Outbound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags: [StockFlows]
      summary: 출고 삭제
      operationId: deleteOutbound
      responses:
        '200':
          description: 출고 삭제 완료
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /alerts:
    get:
      tags: [Alerts]
      summary: 재고부족 알림 로그 조회
      operationId: listAlerts
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: 알림 로그 목록
        '401':
          $ref: '#/components/responses/Unauthorized'
  /alerts/test:
    post:
      tags: [Alerts]
      summary: 테스트 알림 발송
      operationId: triggerAlertTest
      responses:
        '200':
          description: 테스트 알림 발송 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertSendResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /alerts/custom:
    post:
      tags: [Alerts]
      summary: 사용자 정의 알림 발송
      operationId: sendCustomAlert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendCustomAlertRequest'
      responses:
        '200':
          description: 사용자 정의 알림 발송 결과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertSendResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    PageParam:
      name: page
      in: query
      description: 페이지 번호
      schema:
        type: integer
        minimum: 1
        default: 1
    SizeParam:
      name: size
      in: query
      description: 페이지당 항목 수
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: 인증 필요
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    TooManyRequests:
      description: 요청이 제한되었습니다
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
    Page:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer
      required: [page, size, total]
...
    SendCustomAlertRequest:
      type: object
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 1000
      required: [message]
    AlertDecision:
      type: object
      properties:
        canSend:
          type: boolean
        reason:
          type: string
          enum: [ok, cooldown, quiet_hours]
        nextAttemptAt:
          type: string
          format: date-time
          nullable: true
      required: [canSend, reason]
    Alert:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
          nullable: true
        level:
          type: string
          enum: [info, warn, low, error]
        channel:
          type: string
          enum: [telegram, slack, email]
        message:
          type: string
        retryReason:
          type: string
          nullable: true
        retryAt:
          type: string
          format: date-time
          nullable: true
        sentAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, level, channel, message, createdAt]
    AlertSendResponse:
      type: object
      properties:
        success:
          type: boolean
        decision:
          $ref: '#/components/schemas/AlertDecision'
        alert:
          $ref: '#/components/schemas/Alert'
          nullable: true
      required: [success, decision]
...
    Product:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        specification:
          type: string
          nullable: true
        unit:
          type: string
          default: EA
        safetyStock:
          type: integer
          minimum: 0
          default: 0
        totalIn:
          type: integer
          default: 0
        totalOut:
          type: integer
          default: 0
        totalReturn:
          type: integer
          default: 0
        remain:
          type: integer
          default: 0
        status:
          type: string
          enum: [normal, warn, low]
        disabled:
          type: boolean
          default: false
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, code, name, unit, remain, status]
    Inbound:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        quantity:
          type: integer
          minimum: 1
        dateIn:
          type: string
          format: date
        note:
          type: string
          nullable: true
      required: [productId, quantity, dateIn]
    Outbound:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        quantity:
          type: integer
          minimum: 1
        dateOut:
          type: string
          format: date
        note:
          type: string
          nullable: true
      required: [productId, quantity, dateOut]
...